requirements.txt
numpy==1.19.4
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.22
